+++ reactive-UI/src/app/canvas-wrapper/commands/sticky-notes.command.ts
@@ -6,9 +6,9 @@ import { Subject } from 'rxjs';
 import { Group } from 'konva/lib/Group';
 import { IRect } from 'konva/lib/types';
 import { TextInputCommands } from './text-input.command';
-import { FormModalService } from '../../../utilities/controls/form-modal.service';
-import { areRectanglesIntersecting } from '../../../utilities/mathematicals/collision';
-import { Point } from '../../../utilities/types/Point';
+import { FormModalService } from '../../../ui-utilities/controls/form-modal.service';
+import { areRectanglesIntersecting } from '../../../ui-utilities/mathematicals/collision';
+import { Point } from '../../../ui-utilities/types/Point';
 import { STANDARD_STICKY_NOTE_SIZE } from '../../../configs/size';
 import { InkAttachedToStickyNoteEvent, StickyNoteMovedEvent, StickyNotePastedEvent } from '../../events/drawings/EventQueue';
 import { ToolCompositionService } from '../../services/states/tool-composition.service';
