+++ reactive-UI/src/app/canvas-wrapper/managers/BackgroundLayer.manager.ts
@@ -1,7 +1,7 @@
 import Konva from 'konva';
 import { LineConfig } from 'konva/lib/shapes/Line';
 import { Injectable } from '@angular/core';
-import { Point } from '../../../utilities/types/Point';
+import { Point } from '../../../ui-utilities/types/Point';
 import { PRIMARY_COLOR } from '../../../configs/theme.constants';
 import { KonvaObjectService } from '../services/3rds/konva-object.service';
