+++ reactive-UI/src/app/canvas-wrapper/managers/UserDrawingLayer.manager.ts
@@ -9,8 +9,8 @@ import { Group } from "konva/lib/Group";
 import guid from "guid";
 import { Line, LineConfig } from "konva/lib/shapes/Line";
 import { TextInputCommands } from "../commands/text-input.command";
-import { FormModalService } from "../../../utilities/controls/form-modal.service";
-import { Point } from "../../../utilities/types/Point";
+import { FormModalService } from "../../../ui-utilities/controls/form-modal.service";
+import { Point } from "../../../ui-utilities/types/Point";
 import { PRIMARY_COLOR } from "../../../configs/theme.constants";
 import { ToBaseEvent, BoardedCreatedEvent, ToDrawingEvent, BaseEvent, StickyNoteMovedEvent, GeneralUndoEvent, AbstractEventQueueItem, StickyNotePastedEvent, InkAttachedToStickyNoteEvent, TextAttachedToStickyNoteEvent, PencilUpEvent, TextEnteredEvent } from "../../events/drawings/EventQueue";
 import { EventsCompositionService } from "../../events/drawings/events-composition.service";
