+++ reactive-UI/src/app/canvas-wrapper/text-input-commands-form/text-input-commands-form.component.ts
@@ -4,14 +4,14 @@ import {
   OnDestroy,
   ViewChild,
 } from '@angular/core';
-import { ModalContentComponent } from '../../../utilities/controls/form-modal/IModalContentComponent';
-import { TextEditorComponent } from '../../../utilities/controls/text-editor/text-editor.component';
+import { ModalContentComponent } from '../../../ui-utilities/controls/form-modal/IModalContentComponent';
+import { TextEditorComponent } from '../../../ui-utilities/controls/text-editor/text-editor.component';
 import { TEXT_PREVIEW_CONTAINER } from '../../../configs/html-ids.constants';
 import Konva from 'konva';
 import { TranslateService } from '@ngx-translate/core';
-import { ColorBoardComponent } from '../../../utilities/painting/color-board/color-board.component';
+import { ColorBoardComponent } from '../../../ui-utilities/painting/color-board/color-board.component';
 import { SUPPORTED_COLORS } from '../../../configs/theme.constants';
-import { Dimension } from '../../../utilities/types/Dimension';
+import { Dimension } from '../../../ui-utilities/types/Dimension';
 import { TextInputCommands } from '../commands/text-input.command';
 
 @Component({
