+++ reactive-UI/src/ui-utilities/icons/google-swipe/google-swipe.component.html
