+++ reactive-UI/src/ui-utilities/icons/user/user.component.ts
