+++ reactive-UI/src/ui-utilities/mathematicals/collision.ts
