+++ reactive-UI/src/ui-utilities/painting/color-board/color-board.component.html
