+++ reactive-UI/src/ui-utilities/painting/communication-objects/DrawingObject.fabric.ts
