+++ reactive-UI/src/ui-utilities/painting/tool-selector/tool-selector.component.scss
