+++ reactive-UI/src/ui-utilities/static-component/warning-box/warning-box.component.ts
