+++ reactive-UI/src/ui-utilities/types/type-simple-changes.ts
